feat: assign sky fibers using sky table in targetdb
note: because of the current non-homogenized way of assigning `obj_id` in `sky` table, there is a possibility to have multiple `(obj_id, cat_id)` pairs in a `pfsDesign`. I used `sky_id` which is uniquely assigned in the table, but we need to come up with an idea to a coherent way of naming scheme for `obj_id`.